// src/interfaces/bo.ts
export interface IArma {
  ID_ARMA: string
  ID_DIMENSAO_ARMA: string
  NM_DIMENSAO_ARMA?: string
  ID_ACABAMENTO_ARMA: string
  NM_ACABAMENTO_ARMA?: string
  ID_SISTEMA_ARMA: string
  NM_SISTEMA_ARMA?: string
  ID_CORONHA_ARMA: string
  NM_CORONHA_ARMA?: string
  CD_IDENT_ARMA?: string
  VL_CALIBRE: string
  VL_QTD_CANOS: string
  QTD_MUNICAO: string
}

export interface IVeiculo {
  ID_VEICULO?: string
  CD_PLACA: string
  CD_CHASSI?: string
  CD_RENAVAM?: string
  CD_ANO_FABRICACAO?: string
  CD_ANO_MODELO?: string
  FL_SEGURADORA: string
  DS_SEGURO: string
  DS_RASTREADOR: string
  DS_TACOGRAFO: string
  ID_TIPO_COMBUSTIVEL: string
  NM_TIPO_COMBUSTIVEL?: string
}

export interface ICelular {
  ID_CELULAR: string
  CEL_IMEI1?: string
  CEL_IMEI2?: string
  CEL_IMEI3?: string
  CEL_IMEI4?: string
  CEL_IMEI_JUSTIFICATIVA?: string
}

export interface IObjeto {
  ID_OBJETO: string
  ID_TIPO_ENV_OBJETO: string
  ID_COR_OBJETO: string
  NM_COR_OBJETO: string
  ID_TIPO_OBJETO: string
  ID_CATEGORIA: string
  ID_MARCA: string
  ID_MODELO: string
  NUM_SERIE: string
  APREENDIDO: string
  ID_MOEDA: string
  VL_VALOR?: string
  ID_UNIDADE_MEDIDA: string
  QTD_OBJETO: string
  NOME_ENVOLVIDO: string
  ID_ENVOLVIDO: string
  NM_TIPO_ENV_OBJETO: string
  ID_OBJ_ARMA: IArma | null
  ID_VEICULO: IVeiculo | null
  ID_CELULAR: ICelular | null
  NM_TIPO_OBJETO: string
  NM_CATEGORIA: string
  NM_MOEDA: string
  ID_FORMA_APRES: string
  NM_FOMRA_APRES: string
  NM_UNIDADE_MEDIDA: string
  NM_MODELO: string
  NM_MARCA: string
  ID_PAIS: string
  NM_PAIS: string
  FOTO: IFoto | null

  CD_PLACA?: string
  CD_CHASSI?: string
  CD_RENAVAM?: string
  CD_ANO_FABRICACAO?: string
  CD_ANO_MODELO?: string
  ID_TIPO_COMBUSTIVEL?: string

  CEL_IMEI1?: string
  CEL_IMEI2?: string
  CEL_IMEI3?: string
  CEL_IMEI4?: string
  CEL_IMEI_JUSTIFICATIVA?: string

  ID_DIMENSAO_ARMA?: string
  ID_ACABAMENTO_ARMA?: string
  ID_SISTEMA_ARMA?: string
  ID_CORONHA_ARMA?: string
  CD_IDENT_ARMA?: string
  VL_CALIBRE?: string
  VL_QTD_CANOS?: string
  QTD_MUNICAO?: string
}
// ...
export interface IFoto {
  ID: string
  ID_BO: string
  ID_TIPO_FOTO: string
  ID_OBJETO: string
  DESCRICAO: string
  FOTO: string
}